Add Maven dependency resolution to the CLI
This commit adds the ability to resolve dependencies in the Smithy CLI.
This allows end users to author models, generate code, and other
activities from the Smithy CLI without needing to use tools like Gradle
or Maven. This lowers the up-front learning curve of most Smithy use
cases to author models and generate code.

Writing Java library code, model build plugins, or transforms is out of
scope of the Smithy CLI. Such use cases require the use of tools like
Gradle or Maven.

To support dependency resolution in the CLI, this commit also introduces
the ``--version`` option to display the current version.

The ``smithy clean`` command was added since it was very useful to test
dependency resolution by removing the classpath-cache file to force
them to be re-resolved.

Finally, using the Smithy CLI directly means that tools like Gradle aren't
automatically detecting the model directory and adding sources. To account
for this, the `sources` property was added to smithy-build.json. It acts
like `imports`, but the models contained in `sources` are considered
part of the model being built rather than just a dependeny (i.e., the
sources plugin will contain these models).

This commit also prints out more useful exceptions when --stacktrace is
on since wrapping things in a DependencyRunnably obscured the underlying
issue.

To actually test maven resolution, this commit also adds integration
tests for the CLI (./gradlew :smithy-cli:integ).