Refactor dependency behavior and coloring
Dependency behavior and forcing color/no_color is now
done only through environment variables. Expanded various
features of the CLI, like how environment variables are handled.
Removed the deprecated method to set a global stdout for the CLI.

Printing from the CLI previously often resulted in hard to read
output because messages were flushed on each new line rather than
as cohesive sections of text. This change updates the CLI to use a
PrintWriter instead, with a buffer, so that large swaths of text are
kept together. Additionally, a new API for creating grouped sections
of text from a CliPrinter were added that include styling methods.

I also updated checkstyle to allow for empty methods with braces on
the same line to accomodate some of the implementations I was added.
I then went and made similar changes to classes in the CLI package
to use this approach.